+++ unlock-app/src/components/interface/locks/List/elements/LockCard.tsx
@@ -13,16 +13,18 @@ import { FiKey as KeyIcon } from 'react-icons/fi'
 import { IconType } from 'react-icons'
 import Link from 'next/link'
 import { Lock } from '~/unlockTypes'
-import { UNLIMITED_KEYS_DURATION } from '~/constants'
+import { DEFAULT_USER_ACCOUNT_ADDRESS, MAX_UINT } from '~/constants'
 import Duration from '~/components/helpers/Duration'
 import { CryptoIcon } from '../../elements/KeyPrice'
 import { IconModal } from '../../Manage/elements/LockIcon'
 import { ImFilePicture as PictureFile } from 'react-icons/im'
+import { useWeb3Service } from '~/utils/withWeb3Service'
+import { useQueries } from '@tanstack/react-query'
+import { ethers } from 'ethers'
 
 interface LockCardProps {
-  lock: Lock
-  network: number
-  isLoading?: boolean
+  lock: any
+  network: string
 }
 
 interface DetailProps {
@@ -30,13 +32,24 @@ interface DetailProps {
   value?: string | React.ReactNode
   prepend?: React.ReactNode
   icon?: IconType
+  isLoading?: boolean
 }
 
 interface LockIconProps {
   lock: Lock
 }
 
-const Detail = ({ label, prepend, icon, value = '-' }: DetailProps) => {
+const DetailPlaceholder = () => {
+  return <div className="w-8 h-4 animate-pulse bg-slate-200"></div>
+}
+
+const Detail = ({
+  label,
+  prepend,
+  icon,
+  value = '-',
+  isLoading,
+}: DetailProps) => {
   return (
     <div className="flex flex-col gap-1">
       <div className="flex items-center gap-1">
@@ -46,14 +59,18 @@ const Detail = ({ label, prepend, icon, value = '-' }: DetailProps) => {
       <div className="flex items-center gap-2">
         {prepend && <div>{prepend}</div>}
         <Tooltip tip={value} label={label} side="bottom">
-          <span className="text-lg font-bold truncate">{value}</span>
+          {isLoading ? (
+            <DetailPlaceholder />
+          ) : (
+            <span className="text-lg font-bold truncate">{value}</span>
+          )}
         </Tooltip>
       </div>
     </div>
   )
 }
 
-export const LockCardPlaceholder = () => {
+export const LocksByNetworkPlaceholder = () => {
   const DetailPlaceholder = () => {
     return (
       <div className="flex flex-col gap-1">
@@ -65,29 +82,47 @@ export const LockCardPlaceholder = () => {
       </div>
     )
   }
-  return (
-    <div className="flex items-center px-12 py-4 bg-white md:h-24 rounded-2xl">
-      <div className="grid items-center justify-between w-full grid-cols-1 gap-4 md:grid-cols-7">
-        <div className="flex gap-7 md:gap-3 md:col-span-3">
-          <div className="rounded-full bg-slate-200 animate-pulse h-14 w-14"></div>
-          <div className="flex flex-col gap-2">
-            <div className="h-6 w-52 animate-pulse bg-slate-200"></div>
-            <div className="flex items-center gap-3">
-              <div className="w-32 h-4 animate-pulse bg-slate-200"></div>
-              <div className="w-4 h-4 animate-pulse bg-slate-200"></div>
-              <div className="w-4 h-4 animate-pulse bg-slate-200"></div>
+
+  const LockCardPlaceHolder = () => {
+    return (
+      <div className="flex items-center px-12 py-4 bg-white md:h-24 rounded-2xl">
+        <div className="grid items-center justify-between w-full grid-cols-1 gap-4 md:grid-cols-7">
+          <div className="flex gap-7 md:gap-3 md:col-span-3">
+            <div className="rounded-full bg-slate-200 animate-pulse h-14 w-14"></div>
+            <div className="flex flex-col gap-2">
+              <div className="h-6 w-52 animate-pulse bg-slate-200"></div>
+              <div className="flex items-center gap-3">
+                <div className="w-32 h-4 animate-pulse bg-slate-200"></div>
+                <div className="w-4 h-4 animate-pulse bg-slate-200"></div>
+                <div className="w-4 h-4 animate-pulse bg-slate-200"></div>
+              </div>
             </div>
           </div>
+          <div className="flex md:col-span-3 gap-14">
+            <DetailPlaceholder />
+            <DetailPlaceholder />
+            <DetailPlaceholder />
+          </div>
+          <div className="flex justify-between gap-2 md:col-span-1 md:ml-auto">
+            <div className="w-40 h-6 animate-pulse bg-slate-200 md:hidden"></div>
+            <div className="w-6 h-6 animate-pulse bg-slate-200"></div>
+          </div>
         </div>
-        <div className="flex md:col-span-3 gap-14">
-          <DetailPlaceholder />
-          <DetailPlaceholder />
-          <DetailPlaceholder />
-        </div>
-        <div className="flex justify-between gap-2 md:col-span-1 md:ml-auto">
-          <div className="w-40 h-6 animate-pulse bg-slate-200 md:hidden"></div>
-          <div className="w-6 h-6 animate-pulse bg-slate-200"></div>
-        </div>
+      </div>
+    )
+  }
+
+  const NetworkNamePlaceholder = () => {
+    return <div className="w-56 h-6 animate-pulse bg-slate-200"></div>
+  }
+
+  return (
+    <div className="flex flex-col gap-4">
+      <NetworkNamePlaceholder />
+      <div className="flex flex-col gap-6">
+        <LockCardPlaceHolder />
+        <LockCardPlaceHolder />
+        <LockCardPlaceHolder />
       </div>
     </div>
   )
@@ -140,18 +175,55 @@ const LockIcon = ({ lock }: LockIconProps) => {
   )
 }
 
-export const LockCard = ({ lock, network, isLoading }: LockCardProps) => {
+export const LockCard = ({ lock, network }: LockCardProps) => {
   const { networks } = useConfig()
-  const lockAddress = lock.address
+  const web3service = useWeb3Service()
+  const tokenAddress = lock?.tokenAddress
+  const lockAddress = lock?.address
   const [isCopied, setCopied] = useClipboard(lockAddress, {
     successDuration: 2000,
   })
-
   const { explorer, baseCurrencySymbol } = networks?.[network] ?? {}
 
   const explorerUrl = explorer?.urls?.address(lockAddress) || '#'
 
-  const symbol = (lock as any)?.currencySymbol ?? baseCurrencySymbol
+  const getBalance = async (
+    address: string,
+    chainId: number,
+    tokenAddress: string
+  ) => {
+    return await web3service.getAddressBalance(
+      address,
+      chainId,
+      tokenAddress === DEFAULT_USER_ACCOUNT_ADDRESS ? undefined : tokenAddress
+    )
+  }
+
+  const getSymbol = async () => {
+    return await web3service.getTokenSymbol(
+      tokenAddress,
+      parseInt(network!, 10)
+    )
+  }
+
+  const [
+    { isLoading: loadingBalance, data: balance },
+    { isLoading: loadingSymbol, data: tokenSymbol },
+  ] = useQueries({
+    queries: [
+      {
+        queryKey: ['getBalance', lockAddress, network, tokenAddress],
+        queryFn: async () =>
+          await getBalance(lockAddress, parseInt(network, 10), tokenAddress),
+      },
+      {
+        queryKey: ['getSymbol', lockAddress, network, tokenAddress],
+        queryFn: async () => await getSymbol(),
+      },
+    ],
+  })
+
+  const symbol = tokenSymbol ?? baseCurrencySymbol
 
   useEffect(() => {
     if (!isCopied) return
@@ -160,14 +232,15 @@ export const LockCard = ({ lock, network, isLoading }: LockCardProps) => {
 
   const lockUrl = `/locks/lock?address=${lockAddress}&network=${network}`
 
-  if (isLoading) return <LockCardPlaceholder />
-
   const duration =
-    lock.expirationDuration === UNLIMITED_KEYS_DURATION ? (
+    lock?.expirationDuration === MAX_UINT ? (
       'Unlimited'
     ) : (
-      <Duration seconds={lock.expirationDuration} />
+      <Duration seconds={lock?.expirationDuration} />
     )
+
+  const keyPrice = ethers.utils.formatEther(lock?.price)
+  const isLoading = loadingBalance || loadingSymbol
 
   return (
     <>
@@ -200,22 +273,20 @@ export const LockCard = ({ lock, network, isLoading }: LockCardProps) => {
           <div className="grid items-center grid-cols-4 gap-3 md:col-span-3 md:gap-14">
             <Detail
               label="Price"
-              value={lock.keyPrice}
+              value={keyPrice}
               icon={TagIcon}
               prepend={<CryptoIcon symbol={symbol} size={25} />}
+              isLoading={isLoading}
             />
             <Detail
               label="Balance"
-              value={lock.balance}
+              value={balance}
               icon={TagIcon}
               prepend={<CryptoIcon symbol={symbol} size={25} />}
+              isLoading={isLoading}
             />
             <Detail label="Key Duration" value={duration} icon={TimeIcon} />
-            <Detail
-              label="Key Sold"
-              value={lock.outstandingKeys?.toString()}
-              icon={KeyIcon}
-            />
+            <Detail label="Key Sold" value={lock?.totalKeys} icon={KeyIcon} />
           </div>
           <div className="md:ml-auto md:col-span-1">
             <Link href={lockUrl} aria-label="arrow right">